pyproject.toml
[tool.poetry]
name = "ibpy"
version = "0.1.0"
description = ""
authors = ["rishabh <connect@rishabhm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
polars = "^0.18.15"
pyarrow = "^13.0.0"
websockets = "^11.0.3"
httpx = "^0.24.1"
python-dateutil = "^2.8.2"
types-python-dateutil = "^2.8.19.14"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
ruff = "^0.0.285"
black = "^23.7.0"
pre-commit = "^3.3.3"
isort = "^5.12.0"
datamodel-code-generator = "^0.21.4"

[tool.pyright]
reportGeneralTypeIssues = false
reportPrivateUsage = false
reportMissingTypeStubs = false
reportFunctionMemberAccess = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnusedFunction = false

[tool.mypy]
plugins = ["pydantic.mypy"]
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
implicit_reexport = false
disallow_untyped_defs = true
follow_imports = 'silent'

[tool.ruff]
line-length = 120
target-version = "py311"
ignore = ["E741"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
